refactor(lock): moved functions in own module
This setup fits better in to the modularity we already showed when
introducing convenience types in the `filename` and `fd` modules
respectively.

Functions, like `lock` and `unlock` now live in their own `functions`
module, allowing clients to more easily use these as well.

* `functions::Error` now supports a `WouldBlock` variant which clearly
   indicates why a non-blocking lock attempt failed.
* `unlock()` return type was adjusted to reflect the changed `Error`
   enumeration.
